+++ features/support/tag_helpers.rb
@@ -14,6 +14,18 @@ def create_remote_tag tag_name
   end
 end
 
+# Creates a tag with the given name in the remote repo that is not on a branch
+def create_standalone_remote_tag tag_name
+  in_secondary_repository do
+    run 'git checkout -b temp'
+    run 'touch a.txt'
+    run 'git add -A'
+    run 'git commit -m "temp"'
+    run "git tag -a #{tag_name} -m '#{tag_name}'"
+    run 'git push --tags'
+  end
+end
+
 
 # Returns the names of all unpushed tags.
 def unpushed_tags