+++ src/cmd/hack.go
@@ -2,6 +2,7 @@ package cmd
 
 import (
 	"github.com/Originate/git-town/src/git"
+	"github.com/Originate/git-town/src/prompt"
 	"github.com/Originate/git-town/src/script"
 	"github.com/Originate/git-town/src/steps"
 	"github.com/Originate/git-town/src/util"
@@ -9,9 +10,7 @@ import (
 	"github.com/spf13/cobra"
 )
 
-type hackConfig struct {
-	TargetBranch string
-}
+var promptForParent bool
 
 var hackCmd = &cobra.Command{
 	Use:   "hack <branch>",
@@ -30,7 +29,7 @@ the main branch is synced with its upstream counterpart.
 This can be disabled by toggling the "new-branch-push-flag" configuration.`,
 	Run: func(cmd *cobra.Command, args []string) {
 		config := getHackConfig(args)
-		stepList := getHackStepList(config)
+		stepList := getAppendStepList(config)
 		runState := steps.NewRunState("hack", stepList)
 		steps.Run(runState)
 	},
@@ -43,8 +42,18 @@ This can be disabled by toggling the "new-branch-push-flag" configuration.`,
 	},
 }
 
-func getHackConfig(args []string) (result hackConfig) {
+func getParentBranch(targetBranch string) string {
+	if promptForParent {
+		parentBranch := prompt.AskForBranchParent(targetBranch, git.GetMainBranch())
+		prompt.EnsureKnowsParentBranches([]string{parentBranch})
+		return parentBranch
+	}
+	return git.GetMainBranch()
+}
+
+func getHackConfig(args []string) (result appendConfig) {
 	result.TargetBranch = args[0]
+	result.ParentBranch = getParentBranch(result.TargetBranch)
 	if git.HasRemote("origin") && !git.IsOffline() {
 		script.Fetch()
 	}
@@ -52,17 +61,7 @@ func getHackConfig(args []string) (result hackConfig) {
 	return
 }
 
-func getHackStepList(config hackConfig) (result steps.StepList) {
-	mainBranchName := git.GetMainBranch()
-	result.AppendList(steps.GetSyncBranchSteps(mainBranchName, true))
-	result.Append(&steps.CreateAndCheckoutBranchStep{BranchName: config.TargetBranch, ParentBranchName: mainBranchName})
-	if git.HasRemote("origin") && git.ShouldNewBranchPush() && !git.IsOffline() {
-		result.Append(&steps.CreateTrackingBranchStep{BranchName: config.TargetBranch})
-	}
-	result.Wrap(steps.WrapOptions{RunInGitRoot: true, StashOpenChanges: true})
-	return
-}
-
 func init() {
+	hackCmd.Flags().BoolVarP(&promptForParent, "prompt", "p", false, "Prompt for the parent branch")
 	RootCmd.AddCommand(hackCmd)
 }