+++ RELEASE_NOTES.md
@@ -2,6 +2,18 @@
 
 ## Unreleased
 
+## 7.2.0 (2018-06-01)
+
+#### New Features
+
+* `git town config`: print perennial branch trees
+* `git town hack`: add `-p` option which prompts for the parent branch instead of using the main development branch
+* when fetching the origin repository, fetch tags that are not attached to pulled commits
+
+#### Changes
+
+* when fetching the upstream repository, only fetch the main branch
+
 ## 7.1.1 (2018-04-09)
 
 #### Bug Fixes